Escape single quotes while translating dropped Win32 paths
When file/folder is dropped to the terminal, its path is translated and
quoted with a pair of single quotes if the profile is of the WSL.

However, the terminal control did not escape single quotes (allowed in
the Win32 subsystem) that need escapes when translated.

On WSL profiles, it causes incorrect path to be pasted when the path
contains one or more single quotes (see Issue #18006 for an example).

With this commit, the terminal control escapes a single quote with a
valid escape sequence `'\''` (finish quote, print a single quote then
begin quote again) when the WSL path translation is required.